05-monitoring/5.1_monitoring_ground_truth/requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9